refactor: LostItemStatus 관련 DTO 및 컨트롤러 코드 구조 개선
