+++ backend/src/main/java/com/daedan/festabook/lostitem/controller/LostItemController.java
@@ -4,6 +4,8 @@ import com.daedan.festabook.global.argumentresolver.FestivalId;
 import com.daedan.festabook.lostitem.dto.LostItemRequest;
 import com.daedan.festabook.lostitem.dto.LostItemResponse;
 import com.daedan.festabook.lostitem.dto.LostItemResponses;
+import com.daedan.festabook.lostitem.dto.LostItemStatusUpdateRequest;
+import com.daedan.festabook.lostitem.dto.LostItemStatusUpdateResponse;
 import com.daedan.festabook.lostitem.dto.LostItemUpdateResponse;
 import com.daedan.festabook.lostitem.service.LostItemService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -77,7 +79,7 @@ class LostItemController {
     })
     public LostItemStatusUpdateResponse updateLostItemStatus(
             @PathVariable Long lostItemId,
-            @RequestBody LostItemStatusRequest request
+            @RequestBody LostItemStatusUpdateRequest request
     ) {
         return lostItemService.updateLostItemStatus(lostItemId, request);
     }
