+++ backend/src/main/java/com/daedan/festabook/lostitem/dto/LostItemStatusUpdateRequest.java
@@ -1,8 +1,8 @@
-package com.daedan.festabook.lostitem.controller;
+package com.daedan.festabook.lostitem.dto;
 
 import com.daedan.festabook.lostitem.domain.PickupStatus;
 
-public record LostItemStatusRequest(
+public record LostItemStatusUpdateRequest(
         PickupStatus status
 ) {
 }
