+++ backend/src/test/java/com/daedan/festabook/lostitem/dto/LostItemStatusUpdateRequestFixture.java
@@ -1,22 +1,21 @@
 package com.daedan.festabook.lostitem.dto;
 
-import com.daedan.festabook.lostitem.controller.LostItemStatusRequest;
 import com.daedan.festabook.lostitem.domain.PickupStatus;
 
 public class LostItemStatusUpdateRequestFixture {
 
     private static final PickupStatus DEFAULT_STATUS = PickupStatus.PENDING;
 
-    public static LostItemStatusRequest create() {
+    public static LostItemStatusUpdateRequest create() {
         return create(
                 DEFAULT_STATUS
         );
     }
 
-    public static LostItemStatusRequest create(
+    public static LostItemStatusUpdateRequest create(
             PickupStatus status
     ) {
-        return new LostItemStatusRequest(
+        return new LostItemStatusUpdateRequest(
                 status
         );
     }
